feat(config): restructure recording module with new service architecture and UI components
- Refactored recording service into modular components (`recording/`) for better maintainability
- Introduced dedicated modules for playback, processing, and metadata handling
- Replaced monolithic `recording.rs` with specialized submodules:
- `merge.rs` for file merging logic
- `playback.rs` for HLS generation and streaming
- `service.rs` as core service implementation
- Updated frontend components to match new backend structure:
- Replaced `RecordingsOverview.svelte` with granular sections
- Added `RecordingActions`, `RecordingBadges`, and filtering components
- Updated route handlers for recordings in `/src/routes/recordings/`
- Maintained backward compatibility for YouTube and Twitch recording flows
- Preserved existing error handling patterns and validation logic
